Add DarkEdif::SetDataBreakpoint, clean up logging
SetDataBreakpoint is meant for debug only. Only a debugger can intercept when the debug memory is accessed, so adding it otherwise is asking your app to crash when it's reached.
Generally you can just add a data breakpoint yourself in a debugger, e.g. by Breakpoint pane > New > Data Breakpoint while attached, or doing a line breakpoint and editing its conditions. However, this has to be redone per debug session as the address of the variable changes.
Setting it via runtime allows the actual addresses to differ in a non-deterministic way (i.e. different control flow to where the data is allocated), and still be monitored.

These are hard to add at runtime due to sandboxing Apple does, and the fact Android is Java-based, and the fact the monitoring is thread-specific in Windows, etc. So debugger set is usually the route.